style: update chat demo
